Split proxy policy and audit helpers
